ai-threat-detection/requirements.txt
# Core dependencies
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0

# Explainability
shap==0.42.1

# Backend
flask==2.3.2
flask-cors==4.0.0

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.6

# Streaming
kafka-python==2.0.2

# Utilities
requests==2.31.0
pyyaml==6.0
